MA-16596-5 Enable nxp bluetooth on i.MX devices
Update bluetooth hal to 008.013 version, which is
released with mxm5x17283 release.

Reviewed-on: http://androidsource.ap.freescale.net/project/17328

+++ fw_loader_io.h
@@ -44,7 +44,7 @@ extern int8 fw_upload_ComWriteChars(int32 mchar_fd, uint8 *pChBuffer, uint32 uiL
 extern int32 fw_upload_ComReadChars(int32 mchar_fd, uint8 *pChBuffer, uint32 uiCount);
 extern void fw_upload_DelayInMs(uint32 uiMs);
 extern int32 fw_upload_ComGetCTS(int32 mchar_fd);
-extern double fw_upload_GetTime(void);
+extern uint64 fw_upload_GetTime(void);
 extern int32 fw_upload_GetBufferSize(int32 mchar_fd);
 #endif  // _FW_LOADER_IO_LINUX_H
